svg/painting/parsing/stroke-dasharray-valid.svg

<svg xmlns="http://www.w3.org/2000/svg" xmlns:h="http://www.w3.org/1999/xhtml" width="800" height="800"><title>SVG Painting: parsing stroke-dasharray with valid values</title><metadata><h:link href="https://svgwg.org/svg2-draft/painting.html#StrokeDasharrayProperty" rel="help"/><h:meta content="stroke-dasharray supports the full grammar 'none | dasharray'" name="assert"/></metadata><h:script src="/resources/testharness.js"/><h:script src="/resources/testharnessreport.js"/><h:script src="/css/support/parsing-testcommon.js"/><script><![CDATA[

test_valid_value("stroke-dasharray", "none");

// dasharray = [ length-percentage | number ]#*
test_valid_value("stroke-dasharray", "10px");
test_valid_value("stroke-dasharray", "20%");
test_valid_value("stroke-dasharray", "calc(2em + 3ex)");

test_valid_value("stroke-dasharray", "10pt 20% 30pc 40in", "10pt, 20%, 30pc, 40in");
test_valid_value("stroke-dasharray", "10vmin, 20vmax, 30em, 40ex");
test_valid_value("stroke-dasharray", "0, 5", ["0, 5", "0px, 5px"]); // Edge/Safari serialize numbers as lengths.

  ]]></script></svg>
